pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reader-sync"
version = "0.1.0"
description = "Local HTML to Readwise Reader synchronization CLI"
authors = [{name = "niceday"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer[all]>=0.12",
  "httpx[http2]>=0.27",
  "aiolimiter>=1.1",
  "beautifulsoup4>=4.12",
  "lxml>=5.2",
  "PyYAML>=6.0",
  "python-dotenv>=1.0",
  "filelock>=3.14",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-mock>=3.14",
  "anyio>=4.2"
]

[project.scripts]
rw-sync = "reader_sync.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
